src/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI/ML libraries
langchain==0.1.0
langchain-core==0.1.0
langchain-openai==0.0.2
openai==1.3.0

# HTTP and streaming
aiohttp==3.9.0
sse-starlette==1.6.5
websockets==12.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3

# JSON processing
orjson==3.10.14

# RunPod (for compatibility)
runpod